Add optional "CA certificate" field to encryption/templates/peap https://bugs.launchpad.net/wicd/+bug/1317217
It seems wicd does not support configuring a CA certificate when
connecting using PEAP-GTC. This is crucial to ensure the server
we're sending our login information to really is the server it
says it is.

The fix seems simple enough: just add a new optional
"CA certificate" field to encryption/templates/peap.

I've attached a patch that works for me. I tested an invalid
certificate (which gives "invalid password" when connecting),
a valid certificate (which connects successfully) and no
certificate (which also connects successfully).