Fix incorrect sorting of available wireless networks when using dBm https://bugs.launchpad.net/wicd/+bug/617993
I confirm that the sorting by dBm is not done. This affects all
the GUI (gtk, curses).

I have a patch for that :

1/ Remove the sorting from networking.py, because this is not where
   it should be done, as the "view" (by strenght/by dBm) is not
   managed in this module.

2/ Do it in wicd-daemon.py depending on the signal display type.
   Btw the sorting order depends on the sorting type, reverse
   shall be false for dBm as we want the highest on top.

I attach a patch, but I am not sure how to proceed and if the patch
is done against the correct version, this patch is for the 1.7.4.2
release.

At least on my machine it's working correctly, I get network sorted
appropriately.